+++ src/main/java/roomescape/domain/Reservation.java
@@ -2,7 +2,7 @@ package roomescape.domain;
 
 import java.time.LocalDate;
 import lombok.Getter;
-import roomescape.global.exception.InvalidReservationException;
+import roomescape.global.exception.reservation.InvalidReservationException;
 
 @Getter
 public class Reservation {
