+++ src/main/java/roomescape/global/exception/reservation/InvalidReservationPagingException.java
@@ -1,4 +1,7 @@
-package roomescape.global.exception;
+package roomescape.global.exception.reservation;
+
+import roomescape.global.exception.ErrorCode;
+import roomescape.global.exception.RoomescapeException;
 
 public class InvalidReservationPagingException extends RoomescapeException {
