+++ src/main/java/roomescape/global/exception/reservationtime/ReservationTimeNotFoundException.java
@@ -1,4 +1,7 @@
-package roomescape.global.exception;
+package roomescape.global.exception.reservationtime;
+
+import roomescape.global.exception.ErrorCode;
+import roomescape.global.exception.RoomescapeException;
 
 public class ReservationTimeNotFoundException extends RoomescapeException {
