+++ src/main/java/roomescape/repository/ReservationTimeRepository.java
@@ -11,7 +11,7 @@ import org.springframework.jdbc.core.namedparam.SqlParameterSource;
 import org.springframework.jdbc.core.simple.SimpleJdbcInsert;
 import org.springframework.stereotype.Repository;
 import roomescape.domain.ReservationTime;
-import roomescape.global.exception.ReservationTimeNotFoundException;
+import roomescape.global.exception.reservationtime.ReservationTimeNotFoundException;
 
 @Repository
 public class ReservationTimeRepository {
