+++ src/main/java/roomescape/repository/ThemeRepository.java
@@ -10,7 +10,7 @@ import org.springframework.jdbc.core.namedparam.SqlParameterSource;
 import org.springframework.jdbc.core.simple.SimpleJdbcInsert;
 import org.springframework.stereotype.Repository;
 import roomescape.domain.Theme;
-import roomescape.global.exception.ThemeNotFoundException;
+import roomescape.global.exception.theme.ThemeNotFoundException;
 
 @Repository
 public class ThemeRepository {
