+++ src/main/java/roomescape/service/ReservationService.java
@@ -9,8 +9,8 @@ import roomescape.service.dto.reservation.CreateReservationCommand;
 import roomescape.domain.Reservation;
 import roomescape.domain.ReservationTime;
 import roomescape.domain.Theme;
-import roomescape.global.exception.ReservationTimeNotFoundException;
-import roomescape.global.exception.ThemeNotFoundException;
+import roomescape.global.exception.reservationtime.ReservationTimeNotFoundException;
+import roomescape.global.exception.theme.ThemeNotFoundException;
 import roomescape.repository.ReservationRepository;
 import roomescape.repository.ThemeRepository;
 import roomescape.repository.ReservationTimeRepository;
