+++ src/main/java/roomescape/service/ReservationTimeService.java
@@ -6,8 +6,8 @@ import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 import roomescape.domain.ReservationTime;
 import roomescape.domain.ReservedTimes;
-import roomescape.global.exception.DuplicateReservationTimeException;
-import roomescape.global.exception.ThemeNotFoundException;
+import roomescape.global.exception.reservationtime.DuplicateReservationTimeException;
+import roomescape.global.exception.theme.ThemeNotFoundException;
 import roomescape.repository.ReservationTimeRepository;
 import roomescape.repository.ThemeRepository;
 import roomescape.service.dto.reservationtime.AvailableReservationTimesCondition;
