+++ src/main/java/roomescape/service/dto/reservationtime/AvailableReservationTimesCondition.java
@@ -1,7 +1,7 @@
 package roomescape.service.dto.reservationtime;
 
 import java.time.LocalDate;
-import roomescape.global.exception.InvalidReservationTimeException;
+import roomescape.global.exception.reservationtime.InvalidReservationTimeException;
 
 public record AvailableReservationTimesCondition(
         Long themeId,
