+++ src/main/java/roomescape/service/dto/reservationtime/CreateReservationTimeCommand.java
@@ -1,7 +1,7 @@
 package roomescape.service.dto.reservationtime;
 
 import java.time.LocalTime;
-import roomescape.global.exception.InvalidReservationTimeException;
+import roomescape.global.exception.reservationtime.InvalidReservationTimeException;
 
 public record CreateReservationTimeCommand(
         LocalTime startAt
