+++ src/test/java/roomescape/domain/ReservationTest.java
@@ -12,7 +12,7 @@ import org.junit.jupiter.api.Test;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.NullAndEmptySource;
 import org.junit.jupiter.params.provider.ValueSource;
-import roomescape.global.exception.InvalidReservationException;
+import roomescape.global.exception.reservation.InvalidReservationException;
 
 class ReservationTest {
