+++ src/test/java/roomescape/domain/ReservationTimeTest.java
@@ -6,7 +6,7 @@ import static org.assertj.core.api.Assertions.assertThatThrownBy;
 import java.time.LocalTime;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
-import roomescape.global.exception.InvalidReservationTimeException;
+import roomescape.global.exception.reservationtime.InvalidReservationTimeException;
 
 class ReservationTimeTest {
