+++ src/test/java/roomescape/domain/ThemeTest.java
@@ -9,7 +9,7 @@ import org.junit.jupiter.api.Test;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.NullAndEmptySource;
 import org.junit.jupiter.params.provider.ValueSource;
-import roomescape.global.exception.InvalidThemeException;
+import roomescape.global.exception.theme.InvalidThemeException;
 
 class ThemeTest {
