+++ src/test/java/roomescape/util/TestDataInitializer.java
@@ -6,8 +6,8 @@ import org.springframework.stereotype.Component;
 import roomescape.domain.Reservation;
 import roomescape.domain.ReservationTime;
 import roomescape.domain.Theme;
-import roomescape.global.exception.ReservationTimeNotFoundException;
-import roomescape.global.exception.ThemeNotFoundException;
+import roomescape.global.exception.reservationtime.ReservationTimeNotFoundException;
+import roomescape.global.exception.theme.ThemeNotFoundException;
 import roomescape.repository.ReservationRepository;
 import roomescape.repository.ReservationTimeRepository;
 import roomescape.repository.ThemeRepository;
